separate TableWriter and TableReader base classes into separate file
- refactor, splitting hdftableio.py from tableio.py
- fix some imports to be relative
- fix some style warnings
