+++ ctapipe/io/__init__.py
@@ -3,12 +3,10 @@ from .hessioeventsource import HESSIOEventSource
 from .array import get_array_layout
 from .eventsourcefactory import EventSourceFactory, event_source
 from .eventseeker import EventSeeker
-from .hdftableio import (
-    TableReader, TableWriter, HDF5TableReader, HDF5TableWriter
-)
-
+from .tableio import TableWriter, TableReader
+from .hdftableio import  HDF5TableReader, HDF5TableWriter
 
 __all__ = [
-    'TableWriter', 'TableReader', 'HDF5TableWriter', 'HDF5TableReader',
+    'HDF5TableWriter', 'HDF5TableReader',
     'EventSeeker', 'EventSourceFactory', 'EventSource', 'event_source'
 ]
