+++ ctapipe/io/hdftableio.py
@@ -1,6 +1,3 @@
-import re
-from abc import abstractmethod, ABCMeta
-from collections import defaultdict
 from functools import partial
 
 import numpy as np
@@ -9,12 +6,13 @@ from astropy.time import Time
 from astropy.units import Quantity
 
 import ctapipe
-from ctapipe.core import Component, Container
+from . import TableWriter, TableReader
+from ..core import Container
 
-__all__ = ['TableWriter',
-           'TableReader',
-           'HDF5TableWriter',
-           'HDF5TableReader']
+__all__ = [
+    'HDF5TableWriter',
+    'HDF5TableReader'
+]
 
 PYTABLES_TYPE_MAP = {
     'float': tables.Float64Col,
@@ -24,99 +22,7 @@ PYTABLES_TYPE_MAP = {
     'int32': tables.Int32Col,
     'int64': tables.Int64Col,
     'bool': tables.BoolCol,
-    }
-
-
-class TableWriter(Component, metaclass=ABCMeta):
-
-    def __init__(self, parent=None, **kwargs):
-        super().__init__(parent, **kwargs)
-        self._transforms = defaultdict(dict)
-        self._exclusions = defaultdict(list)
-
-    def __enter__(self):
-
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-
-        self.close()
-
-    def exclude(self, table_name, pattern):
-        """
-        Exclude any columns matching the pattern from being written
-
-        Parameters
-        ----------
-        table_name: str
-            name of table on which to apply the exclusion
-        pattern: str
-            regular expression string to match column name
-        """
-        self._exclusions[table_name].append(re.compile(pattern))
-
-    def _is_column_excluded(self, table_name, col_name):
-        for pattern in self._exclusions[table_name]:
-            if pattern.match(col_name):
-                return True
-        return False
-
-    def add_column_transform(self, table_name, col_name, transform):
-        """
-        Add a transformation function for a column. This function will be
-        called on the value in the container before it is written to the
-        output file.
-
-        Parameters
-        ----------
-        table_name: str
-            identifier of table being written
-        col_name: str
-            name of column in the table (or item in the Container)
-        transform: callable
-            function that take a value and returns a new one
-        """
-        self._transforms[table_name][col_name] = transform
-        self.log.debug("Added transform: {}/{} -> {}".format(table_name,
-                                                             col_name,
-                                                             transform))
-
-    @abstractmethod
-    def write(self, table_name, containers):
-        """
-        Write the contents of the given container or containers to a table.
-        The first call to write  will create a schema and initialize the table
-        within the file.
-        The shape of data within the container must not change between calls,
-        since variable-length arrays are not supported.
-
-        Parameters
-        ----------
-        table_name: str
-            name of table to write to
-        container: `ctapipe.core.Container`
-            container to write
-        """
-        pass
-
-    @abstractmethod
-    def open(self, filename):
-
-        pass
-
-    @abstractmethod
-    def close(self):
-
-        pass
-
-    def _apply_col_transform(self, table_name, col_name, value):
-        """
-        apply value transform function if it exists for this column
-        """
-        if col_name in self._transforms[table_name]:
-            tr = self._transforms[table_name][col_name]
-            value = tr(value)
-        return value
+}
 
 
 class HDF5TableWriter(TableWriter):
@@ -141,7 +47,8 @@ class HDF5TableWriter(TableWriter):
 
     Multiple tables may be written at once in a single file, as long as you
     change the table_name attribute to write() to specify which one to write
-    to.
+    to.  To append to existing tables, pass the `mode='a'` option to the
+    constructor.
 
     TODO:
     - ability to write several containers to the same table (appending a
@@ -314,76 +221,6 @@ class HDF5TableWriter(TableWriter):
             self._setup_new_table(table_name, containers)
 
         self._append_row(table_name, containers)
-
-
-class TableReader(Component, metaclass=ABCMeta):
-    """
-    Base class for row-wise table readers. Generally methods that read a
-    full table at once are preferred to this method, since they are faster,
-    but this can be used to re-play a table row by row into a
-    `ctapipe.core.Container` class (the opposite of TableWriter)
-    """
-
-    def __init__(self):
-        super().__init__()
-        self._cols_to_read = defaultdict(list)
-        self._transforms = defaultdict(dict)
-
-    def __enter__(self):
-
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-
-        self.close()
-
-    def add_column_transform(self, table_name, col_name, transform):
-        """
-        Add a transformation function for a column. This function will be
-        called on the value in the container before it is written to the
-        output file.
-
-        Parameters
-        ----------
-        table_name: str
-            identifier of table being written
-        col_name: str
-            name of column in the table (or item in the Container)
-        transform: callable
-            function that take a value and returns a new one
-        """
-        self._transforms[table_name][col_name] = transform
-        self.log.debug("Added transform: {}/{} -> {}".format(table_name,
-                                                             col_name,
-                                                             transform))
-
-    def _apply_col_transform(self, table_name, col_name, value):
-        """
-        apply value transform function if it exists for this column
-        """
-        if col_name in self._transforms[table_name]:
-            tr = self._transforms[table_name][col_name]
-            value = tr(value)
-        return value
-
-    @abstractmethod
-    def read(self, table_name, container):
-        """
-        Returns a generator that reads the next row from the table into the
-        given container.  The generator returns the same container. Note that
-        no containers are copied, the data are overwritten inside.
-        """
-        pass
-
-    @abstractmethod
-    def open(self, filename):
-
-        pass
-
-    @abstractmethod
-    def close(self):
-
-        pass
 
 
 class HDF5TableReader(TableReader):
